Move sct context functions to separate module
